Move customize-gas-modal to its own folder and minor cleanup
