+++ ui/app/components/customize-gas-modal/gas-slider.js
